Replaced double with a decimal data type on multiple variables

+++ Models/Bottles.cs
@@ -20,7 +20,7 @@ namespace TalaGrid.Models
         int bBCId;
 
         [ObservableProperty]
-        double amount;
+        decimal amount;
 
         [ObservableProperty]
         static int adminId;
